requirements.txt
Pillow
python-decouple
djangorestframework
psycopg2-binary
